time reference delta calculation is wrong
time reference is of boottime clock. Currently we are using
uptimemillis() for current time, but that is of a different
clock (monotonic). So this time is going to be wrong. Also
the monotonic clock stops ticking when device is in suspend
mode. So it is not a good choice of clock anyways.

Change-Id: Ia9dd22345329bb8e56492310088f9cab2727e6ce
CRs-Fixed: 2598670