Revised handling of subscription clearing in LCAImpl
1. LCA and HIDL clients behave differently in the way
they need subscriptions to be cleared and setup.
Added some non virtual calls into the LCAImpl layer
which the LCA can call to manage the difference, for example
stopTrackingAndClearSubscriptions which will clear subscriptions
and then stop the session. HIDL clients can continue to call
virtual stopTracking API which by default will not clear subscriptions
now.
2. Add a field to LocAPIStopTrackingReqMsg IPC message
to indicate to hal daemon to clear ths subscriptions at its end or not.
3. Split some of the virtual ILocationAPI functions like startTracking,
updateCallbacks into sync and async counter parts so that all the
messages are handled in LocMsg task for HIDL and LCA clients.
4. Implement the virtual updateTrackingOptions for HIDL clients. This
API is not excercized from LCA clients today.

CRs-Fixed: 3120185

Change-Id: I50781cdb00a9f8cd5c8afa440d6dad4be6764558