Loc api v02: need to register for engine state if sv meas is reged
When sv measurement is subscribed, we need to subscribe for engine
state to determine isFullTracking mode. isFullTracking mode is
using clock reset cnt in the measurement. However, due to quick
session start/stop, the measurement from previous session can
be counted towards next session, and then engine is turned off
and back on, which caused the clock reset and caused the
isFullTracking to be wrong.

Fix is to note down engine off/on while in session, so meas
after engine off/on can be handled properly for isFullTracking
mode.

CRs-fixed: 3620958
Change-Id: I3802983293597f17a916bbf4cffb3351770cae5c